+++ .dovenv/index.js
@@ -10,8 +10,18 @@ export default pigeonposseMonorepoTheme( {
 			themeConfig     : { outline: { level: [ 2, 3 ] } },
 			metaChunk       : true,
 		},
+		styles : { color : {
+			primary   : '#221853',
+			secondary : '#725aa6',
+			terciary  : '#9276b3',
+			fourth    : '#9c7cbc',
+		} },
 		twoslash : false,
 	},
+	predocs : { emoji : {
+		toolkit : '🧰',
+		utils   : '🛠️',
+	} },
 	core : await getWorkspaceConfig( {
 		metaURL  : import.meta.url,
 		path     : '../',