+++ packages/core/src/index.ts
@@ -1,24 +1,42 @@
 
-import { ClippiumData } from './_shared/data'
-import { Finder }       from './_shared/finder'
+import {
+	ClippiumData,
+	InferFlag,
+	InferPosicional,
+} from './_shared/data'
+import { Finder } from './_shared/finder'
 import {
 	Help,
 	HelpConfig,
 } from './help'
 import { ParserData } from './parser'
 
-import type { PartialDeep } from './_shared/types'
-import type { ParserRes }   from './parser'
+import type {
+	Argv,
+	PartialDeep,
+} from './_shared/types'
+import type { ParserRes } from './parser'
 
 type ClippiumConfig = { help?: PartialDeep<HelpConfig> }
-type Argv = string[]
+
+type FnData<C extends ClippiumData> = ParserRes<C>['value'] & { utils : {
+	data       : C
+	config     : ClippiumConfig
+	getHelp    : () => string
+	getVersion : () => string | undefined
+	argv       : Argv
+	parsedArgv : ParserRes<C>['raw']
+} }
 
 export type {
 	ClippiumData,
 	ClippiumConfig,
+	InferFlag,
+	InferPosicional,
 }
 
 export const defineData = <C extends ClippiumData>( data: C ) => data
+export const hiddenBin = ( args: Argv ) => args.slice( 2 )
 
 export class Clippium<C extends ClippiumData> {
 
@@ -35,12 +53,7 @@ export class Clippium<C extends ClippiumData> {
 
 	}
 
-	fn: ( data: ParserRes<C>['value'] & { utils : {
-		getHelp    : () => string
-		getVersion : () => string | undefined
-		argv       : Argv
-		parsedArgv : ParserRes<C>['raw']
-	} } ) => Promise<void> | void = _args => {
+	fn: ( data: FnData<C> ) => Promise<void> | void = _args => {
 
 		// nothing by default
 
@@ -71,6 +84,8 @@ export class Clippium<C extends ClippiumData> {
 		return await this.fn( {
 			...parsed.value,
 			utils : {
+				data       : this.data,
+				config     : this.config,
 				getHelp,
 				getVersion,
 				argv,