+++ packages/presets/config/src/_shared/toml.ts
@@ -1,0 +1,10 @@
+import { parse } from 'smol-toml'
+
+import type { CommonObj } from './_super'
+
+export const getObjectFromTOMLContent = async <Res extends CommonObj = CommonObj>( content: string ) => {
+
+	const r = parse( content ) as Res
+	return r
+
+}