+++ packages/toolkit/src/bin.ts
@@ -1,0 +1,8 @@
+#!/usr/bin/env node
+
+import { hiddenBin } from 'clippium'
+import process       from 'node:process'
+
+import cli from './index'
+
+await cli.run( hiddenBin( process.argv ) )