+++ packages/utils/color/examples/index.ts
@@ -1,0 +1,3 @@
+import color from '../dist/index.js'
+
+console.log( color.inverse.bold( ' Hello ' ) )