+++ packages/toolkit/CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - Add new presets, add toolkit pkg, fix issues
 
+- Updated dependencies []:
+  - @clippium/preset-colored@0.0.2
+  - clippium@0.0.2
+  - @clippium/preset-default@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes