+++ packages/utils/i18n/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @clippium/i18n
 
+## 0.0.2
+
+### Patch Changes
+
+- Add new presets, add toolkit pkg, fix issues
+
 ## 0.0.1
 
 ### Patch Changes